+++ graph/graph_test.go
@@ -1,0 +1,9 @@
+package graph_test
+
+import (
+	"k8s.io/klog/v2"
+)
+
+func init() {
+	klog.InitFlags(nil)
+}