+++ graph/ragged2d_test.go
@@ -30,6 +30,22 @@ func TestRagged2D(t *testing.T) {
 	}, 0)
 }
 
+func TestRagged2D_ReduceSumCols(t *testing.T) {
+	backend := graphtest.BuildTestBackend()
+	// The inputs as fed as parameters, which prevents constant subexpression optimization.
+	// This was used to point out a difference in CPU/GPU versions.
+	output := ExecOnce(backend, func(flat, rowIDs *Node) *Node {
+		r := MakeRagged2D(3, flat, rowIDs)
+		output := r.ReduceSumCols()
+		output.SetLogged("#full Ragged2D.ReduceSumCols")
+		return output
+	},
+		[]float32{1, 3, 5, 7, 11, 13, 17, 19},
+		[]int32{0, 0, 0, 0, 0, 1, 1, 1},
+	)
+	require.Equal(t, []float32{1 + 3 + 5 + 7 + 11, 13 + 17 + 19, 0}, output.Value().([]float32))
+}
+
 func TestRagged2DSoftmax(t *testing.T) {
 	graphtest.RunTestGraphFn(t, "Ragged2DSoftmax", func(g *Graph) (inputs, outputs []*Node) {
 		r := MakeRagged2D(5, Const(g, []float32{1, 3, 5, 7, 11, 13}), Const(g, []int32{0, 0, 0, 1, 1, 3}))
@@ -40,5 +56,4 @@ func TestRagged2DSoftmax(t *testing.T) {
 	}, []any{
 		[]float32{0.015876241, 0.11731043, 0.86681336, 0.01798621, 0.98201376, 1},
 	}, 1e-3)
-
 }