+++ types/tensors/ondevice.go
@@ -219,11 +219,58 @@ func (t *Tensor) lockedInvalidateOnDevice() {
 	})
 }
 
+// OnDeviceClone creates a clone of the tensor t that has backend storage.
+func (t *Tensor) OnDeviceClone(backend backends.Backend, deviceNums ...backends.DeviceNum) *Tensor {
+	if len(deviceNums) == 0 {
+		deviceNums = defaultDeviceNums
+	}
+	if len(deviceNums) > 1 {
+		exceptions.Panicf("Tensor.OnDeviceClone: tensor synchronization across multiple devices not supported yet -- you can create one tensor per device though")
+	}
+	deviceNum := deviceNums[0]
+
+	newT := newTensor(t.Shape())
+	newT.backend = backend
+	var buffer backends.Buffer
+	if backend.HasSharedBuffers() {
+		buffer, newT.sharedFlat = backend.NewSharedBuffer(deviceNum, t.shape)
+		t.ConstFlatData(func(flat any) {
+			reflect.Copy(reflect.ValueOf(newT.sharedFlat), reflect.ValueOf(flat))
+		})
+	} else {
+		t.ConstFlatData(func(flat any) {
+			buffer = newT.backend.BufferFromFlatData(deviceNum, flat, newT.shape)
+		})
+	}
+	d := &onDevice{
+		t:         newT,
+		buffer:    buffer,
+		deviceNum: deviceNum,
+	}
+	newT.onDevices[deviceNum] = d
+	return newT
+}
+
+// Clone creates a clone of the Tensor value with shared backing with the backend -- if it supports --
+// or it falls back to LocalClone.
+func (t *Tensor) Clone() *Tensor {
+	if t.backend == nil || !t.backend.HasSharedBuffers() {
+		return t.LocalClone()
+	}
+	return t.OnDeviceClone(t.backend)
+}
+
 // MaterializeLocal will make sure there is a local storage of the tensor.
 // If there isn't already a local copy, this triggers a transfer from an on-device storage to a local copy.
+//
+// It's a No-op if using shared buffers with the accelerator. If you want to force a local copy,
+// use LocalClone instead.
 func (t *Tensor) MaterializeLocal() {
 	t.mu.Lock()
 	defer t.mu.Unlock()
+	if t.isShared {
+		return
+	}
 	t.lockedMaterializeLocal()
 }
 
@@ -231,6 +278,9 @@ func (t *Tensor) MaterializeLocal() {
 // If there isn't already a local copy, this triggers a transfer from an on-device storage to a local copy.
 //
 // Usually, this is called automatically by all methods that provide Go access to the data (e.g.: Tensor.ConstFlatData).
+//
+// Notice this is not aware of shared buffers: shared buffer handling -- in which case likely there is no
+// need to materialize a local copy -- has to be done by the caller.
 func (t *Tensor) lockedMaterializeLocal() {
 	if t.local != nil && !t.local.IsFinalized() {
 		return